requirements.txt
numpy
pandas
demucs
ffmpeg-python
pretty_midi 
mido 
python-rtmidi 
midi2audio